Modified queries to create db

+++ Queries/CreateDB.sql
@@ -20,7 +20,7 @@ CREATE TABLE [dbo].[Dishes] (
   [name] VARCHAR(100) NOT NULL,
   [description] TEXT,
   [ingredients] TEXT,
-  [price] DECIMAL NOT NULL,
+  [price] FLOAT NOT NULL,
   [stock] INT NOT NULL
 )
 
@@ -28,7 +28,7 @@ CREATE TABLE [dbo].[Drinks] (
   [id] INT NOT NULL PRIMARY KEY IDENTITY,
   [name] VARCHAR(100) NOT NULL,
   [alcoholic] BIT NOT NULL,
-  [price] DECIMAL NOT NULL,
+  [price] FLOAT NOT NULL,
   [stock] INT NOT NULL
 )
 
@@ -58,7 +58,7 @@ CREATE TABLE [dbo].[Orders] (
   [date] DATE NOT NULL,
   [timeOrdering] TIME NOT NULL,
   [timeFinished] TIME NULL,
-  [totalPrice] DECIMAL NOT NULL,
+  [totalPrice] FLOAT NOT NULL,
   [tableId] INT FOREIGN KEY REFERENCES Tables(id) ON DELETE CASCADE NOT NULL,
   [employeeId] INT FOREIGN KEY REFERENCES Employees(id) ON DELETE CASCADE NOT NULL,
 )
